+++ addon/components/toolbar/mark.ts
@@ -1,5 +1,6 @@
 import { action } from '@ember/object';
 import Component from '@glimmer/component';
+import { toggleMarkAddFirst } from '@lblod/ember-rdfa-editor/commands';
 import SayController from '@lblod/ember-rdfa-editor/core/say-controller';
 
 type Args = {
@@ -21,6 +22,7 @@ export default class BoldComponent extends Component<Args> {
 
   @action
   toggle() {
-    this.controller.toggleMark(this.mark);
+    this.controller.focus();
+    this.controller.doCommand(toggleMarkAddFirst(this.mark));
   }
 }